requirements.txt
aiohttp
requests
Pillow
PyPDF2
python-dotenv
colorlog